+++ src/datastore/index.js
@@ -76,6 +76,8 @@ Defaults.prototype.defaultFilter = function (collection, resourceName, params, o
               keep = first ? (attrs[field] <= val) : keep && (attrs[field] <= val);
             } else if (op === 'in') {
               keep = first ? _this.utils.contains(val, attrs[field]) : keep && _this.utils.contains(val, attrs[field]);
+            }  else if (op === 'notIn') {
+              keep = first ? !_this.utils.contains(val, attrs[field]) : keep && !_this.utils.contains(val, attrs[field]);
             } else if (op === '|==') {
               keep = first ? (attrs[field] == val) : keep || (attrs[field] == val);
             } else if (op === '|===') {
@@ -94,6 +96,8 @@ Defaults.prototype.defaultFilter = function (collection, resourceName, params, o
               keep = first ? (attrs[field] <= val) : keep || (attrs[field] <= val);
             } else if (op === '|in') {
               keep = first ? _this.utils.contains(val, attrs[field]) : keep || _this.utils.contains(val, attrs[field]);
+            } else if (op === '|notIn') {
+              keep = first ? !_this.utils.contains(val, attrs[field]) : keep || !_this.utils.contains(val, attrs[field]);
             }
             first = false;
           });
@@ -185,6 +189,8 @@ Defaults.prototype.defaultFilter = function (collection, resourceName, params, o
 Defaults.prototype.baseUrl = '';
 Defaults.prototype.endpoint = '';
 Defaults.prototype.useClass = true;
+Defaults.prototype.keepChangeHistory = false;
+Defaults.prototype.resetHistoryOnInject = true;
 /**
  * @doc property
  * @id DSProvider.properties:defaults.beforeValidate